feat: table data for publications
