+++ public/stats/js/partial-publication-charts.js
