+++ public/stats/partial-publicationCharts.html
@@ -65,7 +65,7 @@
   <script src="https://www.gstatic.com/charts/loader.js"></script>
   <script type="text/javascript" src="/js/common.js"></script>
   <script type="text/javascript" src="/js/helpers/reports.js"></script>
-  <script type="text/javascript" src="./js/publication-charts.js"></script>
+  <script type="text/javascript" src="./js/partial-publication-charts.js"></script>
   <script type="text/javascript" src="/js/pdf-handler.js"></script>
   <script type="text/javascript" src="/js/darkmode.js"></script>
 </body>
