+++ public/stats/partial-visitscharts.html
@@ -63,7 +63,7 @@
 </body>
 <script type="text/javascript" src="/js/pdf-handler.js"></script>
 <script type="text/javascript" src="/js/common.js"></script>
-<script type="text/javascript" src="./js/visitscharts.js"></script>
+<script type="text/javascript" src="./js/partial-visitscharts.js"></script>
 <script type="text/javascript" src="/js/helpers/reports.js"></script>
 <script type="text/javascript" src="/js/darkmode.js"></script>
