+++ public/stats/stats.html
@@ -43,15 +43,22 @@
       </button>
     </a>
   </div>
+  <div class="container" id="pageHeader" style="font-size: 12px; align-content: center; display: flex; align-items: center; justify-content: center">
+    <h2>Publications:</h2>
+  </div>
   <div class="container container-mobile" style="font-size: 12px; align-content: center; display: flex; align-items: center; justify-content: center">
-    <a href="./publicationstats.html">
-      <button class="button-primary button-mobile" id="btnPublications" style="margin: 10; font-size: 22px">
-        Publications
+    <a href="./publicationtables.html">
+      <button class="button-primary button-mobile" id="btnPublicationTables" style="margin: 10; font-size: 22px">
+        Tables
+      </button>
+    </a><a href="./publicationstats.html">
+      <button class="button-primary button-mobile" id="btnPublicationCharts" style="margin: 10; font-size: 22px">
+        Charts
       </button>
     </a>
     <a href="./publication-trends.html?data=articles&resolution=month">
-      <button class="button-primary button-mobile" id="btnPublications" style="margin: 10; font-size: 22px">
-        Publication Trends
+      <button class="button-primary button-mobile" id="btnPublicationTrends" style="margin: 10; font-size: 22px">
+        Trends
       </button>
     </a>
   </div>
